+++ build_runner/lib/src/generate/watch_impl.dart
@@ -231,7 +231,9 @@ class WatchImpl implements BuildState {
 
     Future<BuildResult> doBuild(List<List<AssetChange>> changes) async {
       var build = _build!;
-      _logger..info('${'-' * 72}\n')..info('Starting Build\n');
+      _logger
+        ..info('${'-' * 72}\n')
+        ..info('Starting Build\n');
       var mergedChanges = collectChanges(changes);
 
       _expectedDeletes.clear();