+++ build_vm_compilers/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.0.9-dev
 
+- Update to build version `2.0.0`.
+
 ## 1.0.8
 
 - Update to build_modules version `3.0.0`.